SCAN4NET-332 Extract ExternalIssueTest
