--- its/projects/ProjectUnderTest/TestQualityProfileCustomRoslyn.xml
@@ -1,13 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<profile>
-  <name>ProfileForTestCustomRoslyn</name>
-  <language>cs</language>
-  <rules>
-    <rule>
-      <repositoryKey>csharpsquid</repositoryKey>
-      <key>S1134</key>
-      <priority>MAJOR</priority>
-      <parameters/>
-    </rule>
-  </rules>
-</profile>
