+++ its/src/test/java/com/sonar/it/scanner/msbuild/sonarqube/ExternalIssuesTest.java
@@ -1,0 +1,148 @@
+/*
+ * SonarScanner for .NET
+ * Copyright (C) 2016-2025 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.sonar.it.scanner.msbuild.sonarqube;
+
+import com.sonar.it.scanner.msbuild.utils.ScannerClassifier;
+import com.sonar.it.scanner.msbuild.utils.TestUtils;
+import com.sonar.orchestrator.build.BuildResult;
+import com.sonar.orchestrator.locator.FileLocation;
+import java.io.IOException;
+import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.io.TempDir;
+import org.sonarqube.ws.Issues.Issue;
+
+import static com.sonar.it.scanner.msbuild.sonarqube.Tests.ORCHESTRATOR;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.tuple;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assumptions.assumeFalse;
+
+@ExtendWith(Tests.class)
+class ExternalIssuesTest {
+  @TempDir
+  public Path basePath;
+
+  @Test
+  void checkExternalIssuesVB() throws Exception {
+    String projectKey = "checkExternalIssuesVB";
+    ORCHESTRATOR.getServer().restoreProfile(FileLocation.of("projects/ExternalIssues.VB/TestQualityProfileExternalIssuesVB.xml"));
+    ORCHESTRATOR.getServer().provisionProject(projectKey, "sample");
+    ORCHESTRATOR.getServer().associateProjectToQualityProfile(projectKey, "vbnet", "ProfileForTestExternalIssuesVB");
+
+    Path projectDir = TestUtils.projectDir(basePath, "ExternalIssues.VB");
+    String token = TestUtils.getNewToken(ORCHESTRATOR);
+
+    TestUtils.newScannerBegin(ORCHESTRATOR, projectKey, projectDir, token, ScannerClassifier.NET_FRAMEWORK).execute(ORCHESTRATOR);
+    TestUtils.runMSBuild(ORCHESTRATOR, projectDir, "/t:Rebuild");
+    BuildResult result = TestUtils.executeEndStepAndDumpResults(ORCHESTRATOR, projectDir, projectKey, token);
+
+    assertTrue(result.isSuccess());
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
+    List<String> ruleKeys = issues.stream().map(Issue::getRule).collect(Collectors.toList());
+
+    // The same set of Sonar issues should be reported, regardless of whether
+    // external issues are imported or not
+    assertThat(ruleKeys).containsAll(Arrays.asList(
+      "vbnet:S112",
+      "vbnet:S3385"));
+
+    if (ORCHESTRATOR.getServer().version().isGreaterThanOrEquals(7, 4)) {
+      // if external issues are imported, then there should also be some CodeCracker errors.
+      assertThat(ruleKeys).containsAll(Arrays.asList(
+        "external_roslyn:CC0021",
+        "external_roslyn:CC0062"));
+
+      assertThat(issues).hasSize(4);
+
+    } else {
+      // Not expecting any external issues
+      assertThat(issues).hasSize(2);
+    }
+  }
+
+  @Test
+  void checkExternalIssuesCS() throws Exception {
+    String projectKey = "ExternalIssues.CS";
+    ORCHESTRATOR.getServer().restoreProfile(FileLocation.of("projects/ExternalIssues.CS/TestQualityProfileExternalIssues.xml"));
+    ORCHESTRATOR.getServer().provisionProject(projectKey, "sample");
+    ORCHESTRATOR.getServer().associateProjectToQualityProfile(projectKey, "cs", "ProfileForTestExternalIssues");
+
+    Path projectDir = TestUtils.projectDir(basePath, projectKey);
+    String token = TestUtils.getNewToken(ORCHESTRATOR);
+
+    TestUtils.newScannerBegin(ORCHESTRATOR, projectKey, projectDir, token, ScannerClassifier.NET_FRAMEWORK).execute(ORCHESTRATOR);
+    TestUtils.runMSBuild(ORCHESTRATOR, projectDir, "/t:Rebuild");
+    BuildResult result = TestUtils.executeEndStepAndDumpResults(ORCHESTRATOR, projectDir, projectKey, token);
+
+    assertTrue(result.isSuccess());
+
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
+    List<String> ruleKeys = issues.stream().map(Issue::getRule).collect(Collectors.toList());
+
+    // The same set of Sonar issues should be reported, regardless of whether
+    // external issues are imported or not
+    assertThat(ruleKeys).containsAll(Arrays.asList(
+      "csharpsquid:S125",
+      "csharpsquid:S1134"));
+
+    if (ORCHESTRATOR.getServer().version().isGreaterThanOrEquals(7, 4)) {
+      // if external issues are imported, then there should also be some
+      // Wintellect errors.  However, only file-level issues are imported.
+      assertThat(ruleKeys).containsAll(List.of(
+        "external_roslyn:Wintellect004"));
+
+      assertThat(issues).hasSize(3);
+
+    } else {
+      // Not expecting any external issues
+      assertThat(issues).hasSize(2);
+    }
+  }
+
+  @Test
+  void testIgnoreIssuesDoesNotRemoveSourceGenerator() throws IOException {
+    assumeFalse(TestUtils.getMsBuildPath(ORCHESTRATOR).toString().contains("2017")); // We can't run .NET Core SDK under VS 2017 CI context
+    var projectKey = "IgnoreIssuesDoesNotRemoveSourceGenerator";
+    Path projectDir = TestUtils.projectDir(basePath, projectKey);
+
+    String token = TestUtils.getNewToken(ORCHESTRATOR);
+
+    TestUtils.newScannerBegin(ORCHESTRATOR, projectKey, projectDir, token, ScannerClassifier.NET_FRAMEWORK)
+      .setProperty("sonar.cs.roslyn.ignoreIssues", "true")
+      .execute(ORCHESTRATOR);
+    TestUtils.buildMSBuild(ORCHESTRATOR, projectDir);
+
+    BuildResult result = TestUtils.executeEndStepAndDumpResults(ORCHESTRATOR, projectDir, projectKey, token);
+
+    assertTrue(result.isSuccess());
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
+    assertThat(issues)
+      .extracting(Issue::getRule, Issue::getComponent)
+      .containsExactlyInAnyOrder(
+        tuple("csharpsquid:S1481", "IgnoreIssuesDoesNotRemoveSourceGenerator:ProjectWithSourceGenAndAnalyzer/Program.cs"),
+        tuple("csharpsquid:S1186", "IgnoreIssuesDoesNotRemoveSourceGenerator:ProjectWithSourceGenAndAnalyzer/Program.cs")
+      );
+  }
+}
